medrisk-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart

# PDF
pymupdf

# HTTP + LLM calls
httpx

# OCR + image processing
pillow
pytesseract
rapidocr-onnxruntime
numpy
opencv-python-headless

# env vars from .env
python-dotenv

# ML
joblib
scikit-learn==1.7.2
pandas